[INFRA] Reorganise Header Files
