+++ include/detect_breakends/method_enums.hpp
