+++ include/structures/aligned_segment.hpp
