+++ include/structures/junction.hpp
