+++ src/detect_breakends/junction_detection.cpp
@@ -4,10 +4,10 @@
 #include <seqan3/io/alignment_file/input.hpp>   // SAM/BAM support
 #include <seqan3/io/sequence_file/output.hpp>   // FASTA support
 
-#include "cluster.hpp"                          // for class Cluster
-#include "detect_breakends/aligned_segment.hpp" // for struct AlignedSegment
 #include "detect_breakends/bam_functions.hpp"   // for hasFlag* functions
-#include "junction.hpp"                         // for class Jnction
+#include "structures/aligned_segment.hpp"       // for struct AlignedSegment
+#include "structures/cluster.hpp"               // for class Cluster
+#include "structures/junction.hpp"              // for class Junction
 
 using seqan3::operator""_cigar_op;
 using seqan3::operator""_tag;
